Override for default registration backend and monkey patch for auth module.
Subclassed the DefaultBackend to override the user creation behavior by using the email address
as the username and taking other user-supplied data and adding it to the User model.  Monkey
patched the auth user validation to allow for longer user names to support this.  Also removed the
catchall URL config so that the APPEND_SLASH setting works.